solutions/java/src/main/java/playground/algorithm/SubtreeOfAnotherTree.java
package playground.algorithm;

import playground.algorithm.common.TreeNode;

public class SubtreeOfAnotherTree {

    public boolean isSubtree(TreeNode s, TreeNode t) {
        if (s == null || t == null) {
            return false;
        }
        if (isSame(s, t)) {
            return true;
        }
        return isSubtree(s.left, t) || isSubtree(s.right, t);
    }

    private boolean isSame(TreeNode s, TreeNode t) {
        if (s == null && t == null) {
            return true;
        }
        if (s == null || t == null) {
            return false;
        }
        if (s.val == t.val && isSame(s.left, t.left) && isSame(s.right, t.right)) {
            return true;
        }
        return false;
    }
}
